Updated Article Content Type
I have updated the content type for Articles.
This branch has updated yml files that should cover both the content type and the new paragraphs element Article Content
Be sure to use the branch 59 of the base repository to make sure you have the updated yml and css for the Articles content type and paragraph.

Related to issue #59 in the base repo